ml/requirements.txt
# Minimal package to use dcase2020 task2 baseline.
keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
matplotlib==3.0.3
numpy==1.16.0
PyYAML==5.1
scikit-learn==0.20.2
librosa==0.6.0
audioread==2.1.5
setuptools==41.0.0
tensorflow==1.15.0
tqdm>=4.43.0
pytorch-lightning
dl-cliche
torchsummary
EasyDict
